Payment's statuses reengeneering

+++ web_module/branches/pacts_status_2007_04_13/src/main/com/topcoder/web/tc/controller/legacy/pacts/common/Payment.java
@@ -32,7 +32,7 @@ public class Payment implements PactsConstants, java.io.Serializable {
     private int methodId;
     private String method;
     private String description;
-    private int statusId;
+    private long statusId;
     private String statusDesc;
     private String modifiedDate;
     private String rationale;
@@ -296,11 +296,11 @@ public class Payment implements PactsConstants, java.io.Serializable {
         this.description = description;
     }
 
-    public int getStatusId() {
+    public long getStatusId() {
         return statusId;
     }
 
-    public void setStatusId(int statusId) {
+    public void setStatusId(long statusId) {
         this.statusId = statusId;
     }
